+++ src/main/java/com/emsi/gowithus/service/IConducteurService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface IConducteurService {
     List<Conducteur> getAllAprrouved();
     List<Conducteur> getAllNonApprouved();
-    Conducteur setApprouvedUser(int id);
+    Conducteur setApprouvedUser(Long id);
 }
